# nem_pay/api/openapi.yaml
openapi: 3.0.3
info:
  title: NemPay API
  version: "1.0.0-M1"
  description: >
    NemPay direct-capture gateway (Milestone M1). Hand-maintained and kept in step with the
    handlers in internal/httpapi. Money verbs (confirm/capture/refund) and webhook config are
    added in later tasks. The M2/M4 merchants generate their clients from this file.
servers:
  - url: http://localhost:8080
    description: Local docker-compose

components:
  securitySchemes:
    SecretKey:
      type: http
      scheme: bearer
      description: >
        A merchant secret key (sk_...) as `Authorization: Bearer <key>`. Required for every
        payment-intent route. A publishable key (pk_...) authenticates but is refused (403) on
        these routes.
    PublishableKey:
      type: http
      scheme: bearer
      description: >
        A merchant publishable key (pk_...) as `Authorization: Bearer <key>`. Safe to expose in a
        browser. Used only for /v1/tokens; a secret key is refused (403) there.
    Session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A portal session token (JWT) from /v1/portal/login, as `Authorization: Bearer <jwt>`.
        Accepted on the read endpoints alongside a secret key; never on money-mutating routes.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, code, message]
          properties:
            type:
              type: string
              enum: [invalid_request_error, authentication_error, idempotency_error, not_found_error, api_error]
            code: { type: string, example: invalid_amount }
            message: { type: string }
            param: { type: string }
    PaymentIntent:
      type: object
      properties:
        id: { type: string, format: uuid }
        object: { type: string, enum: [payment_intent] }
        amount:
          type: integer
          format: int64
          description: Amount in minor units (e.g. cents). Always positive.
        currency: { type: string, minLength: 3, maxLength: 3, example: USD }
        status:
          type: string
          enum: [created, requires_confirmation, authorized, captured, settled, failed, refunded, partially_refunded]
        settlement_mode: { type: string, enum: [direct] }
        metadata: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateIntentRequest:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: integer, format: int64, minimum: 1 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        metadata: { type: object, additionalProperties: true }
    IntentList:
      type: object
      properties:
        object: { type: string, enum: [list] }
        data:
          type: array
          items: { $ref: '#/components/schemas/PaymentIntent' }

paths:
  /v1/health:
    get:
      summary: Liveness check (public, no auth)
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }

  /v1/tokens:
    post:
      summary: Tokenize a card (browser-facing, publishable key)
      description: >
        Called from the browser with a publishable key; the card PAN reaches the gateway here (never
        a merchant server) and is exchanged for a single-use payment-method token. The PAN is not
        stored — only brand and last4 are derived. CORS-enabled for the configured merchant origin.
      security: [{ PublishableKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [number, exp_month, exp_year, cvc]
              properties:
                number: { type: string, example: "4242424242424242" }
                exp_month: { type: integer, example: 12 }
                exp_year: { type: integer, example: 2030 }
                cvc: { type: string, example: "123" }
      responses:
        '200':
          description: A payment-method token
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, example: tok_ok }
                  object: { type: string, example: token }
                  card:
                    type: object
                    properties:
                      brand: { type: string, example: visa }
                      last4: { type: string, example: "4242" }
                      exp_month: { type: integer, example: 12 }
                      exp_year: { type: integer, example: 2030 }
        '400': { description: Invalid card fields, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Missing/invalid key, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Secret key used (publishable required), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  /v1/portal/login:
    post:
      summary: Portal login — exchange credentials for a session token (public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: owner@merchant-a.test }
                password: { type: string, example: portal-dev-password }
      responses:
        '200':
          description: A session token bound to the user's merchant
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expires_at: { type: string, format: date-time }
                  merchant:
                    type: object
                    properties:
                      id: { type: string }
                      name: { type: string }
        '401': { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  /v1/balances:
    get:
      summary: Derived balances (by type/kind/currency) for the authenticated merchant
      security: [{ SecretKey: [] }, { Session: [] }]
      responses:
        '200':
          description: Balances derived from the ledger (Σ entries)
          content:
            application/json:
              schema:
                type: object
                properties:
                  balances:
                    type: array
                    items:
                      type: object
                      properties:
                        type: { type: string, example: asset }
                        kind: { type: string, example: platform_cash }
                        currency: { type: string, example: USD }
                        balance: { type: integer, format: int64, description: minor units }

  /v1/webhook_events:
    get:
      summary: Emitted webhook events for the merchant (newest first)
      security: [{ SecretKey: [] }, { Session: [] }]
      parameters:
        - { in: query, name: limit, schema: { type: integer, default: 20, maximum: 100 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Webhook events with delivery status
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook_events:
                    type: array
                    items:
                      type: object
                      properties:
                        event_id: { type: string }
                        event_type: { type: string, example: payment_intent.captured }
                        status: { type: string, example: delivered }
                        attempts: { type: integer }
                        last_error: { type: string, nullable: true }
                        created_at: { type: string, format: date-time }

  /v1/api_keys:
    get:
      summary: The merchant's API keys (secret masked — prefix only)
      security: [{ SecretKey: [] }, { Session: [] }]
      responses:
        '200':
          description: API keys by kind and prefix (never the full key)
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        kind: { type: string, example: secret }
                        prefix: { type: string, example: sk_test_ }
                        masked: { type: string, example: "sk_test_••••" }
                        created_at: { type: string, format: date-time }
                        revoked_at: { type: string, format: date-time, nullable: true }

  /v1/payment_intents/{id}/ledger:
    get:
      summary: The ledger transaction(s)/entries backing one intent
      security: [{ SecretKey: [] }, { Session: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: Transactions with their double-entry entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        kind: { type: string, example: capture }
                        created_at: { type: string, format: date-time }
                        entries:
                          type: array
                          items:
                            type: object
                            properties:
                              account_type: { type: string }
                              account_kind: { type: string }
                              debit: { type: integer, format: int64 }
                              credit: { type: integer, format: int64 }
                              currency: { type: string }
        '404': { description: No such intent for this merchant, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  /v1/payment_intents:
    post:
      summary: Create a payment intent
      security: [{ SecretKey: [] }]
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
          description: Required on every money-mutating POST; retries replay the stored response.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateIntentRequest' }
      responses:
        '200':
          description: Created intent (status = created)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentIntent' }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Missing/invalid key, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Publishable key used, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: A request with this Idempotency-Key is still in flight, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: Idempotency-Key reused with a different body, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    get:
      summary: List payment intents (newest first)
      security: [{ SecretKey: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, default: 20, maximum: 100 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: A list of intents
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntentList' }

  /v1/payment_intents/{id}:
    get:
      summary: Retrieve a payment intent
      security: [{ SecretKey: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: The intent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentIntent' }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }